% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coinbase-package.R
\docType{package}
\name{coinbase-package}
\alias{coinbase}
\alias{coinbase-package}
\title{coinbase: API Wrapper to Coinbase Advanced Trade}
\description{
R API wrapper to the Coinbase Advanced Trade API supporting both synchronous and asynchronous (promise based) operations. Provides R6 classes for market data, spot trading, account management, and US futures (CFM), with helpers for tick-to-OHLCV aggregation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dereckscompany.github.io/coinbase}
  \item \url{https://github.com/dereckscompany/coinbase}
  \item Report bugs at \url{https://github.com/dereckscompany/coinbase/issues}
}

}
\author{
\strong{Maintainer}: Dereck Mezquita \email{dereck@mezquita.io} (\href{https://orcid.org/0000-0002-9307-6762}{ORCID})

}
\keyword{internal}
