% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_client_order_id}
\alias{generate_client_order_id}
\title{Generate a Client Order ID}
\usage{
generate_client_order_id()
}
\value{
Character; a UUID, e.g. \code{"11299b2b-61e3-43e7-b9f7-dee77210bb29"}.
}
\description{
Produces a random RFC 4122 version-4 UUID string for use as the
\code{client_order_id} idempotency key when placing orders.
}
\examples{
generate_client_order_id()

}
