% Generated by roxyassert: do not edit by hand.
\name{roxyassert-generated-asserts}
\alias{roxyassert-generated-asserts}
\alias{assert_type_Accounts}
\alias{assert_type_BestBidAsk}
\alias{assert_type_CancelResults}
\alias{assert_type_CreateOrderAck}
\alias{assert_type_EditOrderAck}
\alias{assert_type_EditPreview}
\alias{assert_type_Fees}
\alias{assert_type_Fills}
\alias{assert_type_FuturesBalance}
\alias{assert_type_FuturesPositions}
\alias{assert_type_FuturesSweeps}
\alias{assert_type_MarginWindow}
\alias{assert_type_Ohlcv}
\alias{assert_type_OrderConfig}
\alias{assert_type_Orders}
\alias{assert_type_PortfolioSummary}
\alias{assert_type_Preview}
\alias{assert_type_Products}
\alias{assert_type_ProductStats}
\alias{assert_type_Stats}
\alias{assert_type_Trades}
\title{Generated assertion helpers}
\usage{
assert_type_Accounts(value)

assert_type_BestBidAsk(value)

assert_type_CancelResults(value)

assert_type_CreateOrderAck(value)

assert_type_EditOrderAck(value)

assert_type_EditPreview(value)

assert_type_Fees(value)

assert_type_Fills(value)

assert_type_FuturesBalance(value)

assert_type_FuturesPositions(value)

assert_type_FuturesSweeps(value)

assert_type_MarginWindow(value)

assert_type_Ohlcv(value)

assert_type_OrderConfig(value)

assert_type_Orders(value)

assert_type_PortfolioSummary(value)

assert_type_Preview(value)

assert_type_Products(value)

assert_type_ProductStats(value)

assert_type_Stats(value)

assert_type_Trades(value)
}
\arguments{
\item{value}{The value or shaped object validated by the assert helper.}
}
\value{
\code{assert_args_*} return \code{invisible(NULL)}; \code{assert_return_*}
return the validated value; \code{assert_type_*} return it invisibly. All are
called for the side effect of signalling an error on a contract violation.
}
\description{
Runtime assertion helpers generated by roxyassert from typed roxygen
documentation and exported via the \verb{@exportassert} tag.
}
\keyword{internal}
